Fold header classes into 3 classes
